Split inst match generator into multiple files
